Use topic device/deviceId; integrate deviceCache
device/deviceId is a more elegant topic for devices to publish to.
Server still needs to know accountId and deviceType in context, so
integrate deviceCache into index.js and alert.js so they don't have to
hit the database to get that information every time. TODO: separate
deviceCache into a separate module that listens for MQTT updates to
device configuration.